fix: compare a Short against a Short in the PCM conversion test (Task22)
assertEquals(0, samples[1]) compares an Int literal with a Short. Kotlin
does not widen either to the other, so both box and the call resolves to
assertEquals(Object, Object) -- where Integer(0) does not equal Short(0),
and the test fails on a value that is correct.

Every other numeric assertion in the suite already spells out .toShort()
or .toByte(); this was the one that did not.

The irony is worth recording rather than quietly fixing: the test that
failed is the one whose whole subject is a silent type mismatch that
neither throws nor fails symmetrically. PcmConversion itself was right.

+++ core/src/test/kotlin/com/saikai/ptt/core/domain/PcmConversionTest.kt
@@ -80,7 +80,7 @@ class PcmConversionTest {
         val samples = ShortArray(2)
         assertEquals(1, PcmConversion.bytesToShorts(byteArrayOf(0x02, 0x01, 0x03), 0, 3, samples, 0))
         assertEquals(0x0102.toShort(), samples[0])
-        assertEquals(0, samples[1])
+        assertEquals(0.toShort(), samples[1])
     }
 
     @Test
